// Denti_Salud/ClientApp/src/app/models/odontologo.ts
export class Odontologo {
  nombre: string = "";
  apellido: string = "";
  fechaNacimiento: Date = undefined;
  tipoIdentificacion: string = "";
  identificacion: string = "";
  correo: string = "";
  genero: string = "";
  direccion: string = "";
  telefono: string = "";
  salario: number = 0;
  cargo: string = "";
}